requirements.txt.txt
python-telegram-bot==13.21
